Developing the Bassline library
The Bassline library is intended to provide convenient structures that
allow for the user to run their computations in a given computational
context. For example, the Maybe structure is provided to allow for a
result to be Just a value or Nothing.

Using these structures, addons like Functors, Applicatives, and Monads
are sitting behind the scenes to give JavaScript a more Haskell-like
feel.