Working out the prototype chain structure
The essence to the prototype structure is to allow for a mechanism which
inserts objects in between an object and the object's prototype. This
creates a longer prototype chain, but allows for convenient code
sharing, and standard structure for introducing "type" classes.

Additionally, after debating between RequireJS and CommonJS, it was
decided to use CommonJS as the main structuring of the modules so that
we have seamless integration with NodeJS.