Enhancing the wrapped objects
Additional specs and functionality has been added for those objects
which are wrapped. For example, the ListW and StringW have new
functionality, along with Sum and a new object Product.

It should also be mentioned that the StringW object is under
consideration for removal. This is because of the complexity surrounding
how a wrapped String should be used in terms of an Applicative Functor
and a Monad. The StringW is useful as a Monoid, so perhaps it will
simply exist for that.