Simplification of bassline
Only the basics were kept, such as the maybe and either functionality,
where the methods for functor, applicative, monad, and folder have been
merged into the objects directly.

The intention of this trimming-down is to provide a very simple release
for use by node.js. However, this means the API is not stable currently,
and may change, but for the time being, the included functionality
should be sufficient for basic use.