Implementing promises for chaining CPS functions
A Promise object has been introduced to allow for the chaining together
of functions that would be otherwise written in continuation passing
style. Promises may be created and then fullfilled with the resulting
value in the callback of functions that accept continuations. The
promise is returned to allow for chaining (binding) of other promises.