Removing the notion of failure in the promise
The functionality of transitioning the promise to the failure state has
been removed because it is not yet necessary. Failure handling can later
be added to the implementation. However, not all is lost because the
caller can locally attempt to resolve an error before invoking fulfill
or they can throw the error to allow for outside exception handling to
process it.